requirements.txt
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.17.0
torch==2.3.0
transformers==4.40.1